requirements.txt
asgiref==3.6.0
boto3==1.26.81
botocore==1.29.81
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==3.2.18
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-storages==1.13.2
djangorestframework==3.14.0
filelock==3.9.0
gunicorn==20.1.0
identify==2.5.20
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
nodeenv==1.7.0
pip-shims==0.3.4
pipenv==2023.2.18
platformdirs==3.1.0
pre-commit==3.1.1
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.20.0
virtualenv-clone==0.5.7
